🛠 Chore: #upgrade_deps converterlib from:v2.0.38 to:v2.0.39

+++ AdGuardSDK/AdGuardSDKTests/Mocks/SafariConfigurationMock.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 final class SafariConfigurationMock: SafariConfigurationProtocol {
-    var iosVersion: Int = 11
+    var iosVersion: Double = 11
 
     var advancedBlockingIsEnabled: Bool = false
